Add type assertions for balance and allowance bigint conversions

+++ components/RedeemForm.tsx
@@ -102,7 +102,7 @@ export default function RedeemForm() {
   // Check if approval is needed
   const needsApproval = useMemo(() => {
     if (allowance === undefined || allowance === null || !amountWei) return false;
-    return BigInt(allowance) < amountWei;
+    return BigInt(allowance as bigint) < amountWei;
   }, [allowance, amountWei]);
 
   // Validate inputs
@@ -110,7 +110,7 @@ export default function RedeemForm() {
     if (!amount || balance === undefined || balance === null || decimals === undefined) return false;
     try {
       const amt = parseUnits(amount, Number(decimals));
-      return amt > 0n && amt <= BigInt(balance);
+      return amt > 0n && amt <= BigInt(balance as bigint);
     } catch {
       return false;
     }
@@ -130,7 +130,7 @@ export default function RedeemForm() {
   // Handle MAX button
   const handleMax = () => {
     if (balance !== undefined && balance !== null && decimals !== undefined) {
-      setAmount(formatUnits(BigInt(balance), Number(decimals)));
+      setAmount(formatUnits(BigInt(balance as bigint), Number(decimals)));
     }
   };
 
@@ -273,7 +273,7 @@ export default function RedeemForm() {
         <div className="flex justify-between items-center">
           <h2 className="text-xl font-semibold">Redeem {symbol || 'plUSD'}</h2>
           <div className="text-sm text-gray-400">
-            Balance: {balance !== undefined && balance !== null && decimals !== undefined ? formatUnits(BigInt(balance), Number(decimals)) : '0'} {symbol || 'plUSD'}
+            Balance: {balance !== undefined && balance !== null && decimals !== undefined ? formatUnits(BigInt(balance as bigint), Number(decimals)) : '0'} {symbol || 'plUSD'}
           </div>
         </div>
